app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Projeto Django + DRF"
authors = ["Seu Nome <seu@email.com>"]
packages = []   # <- importante para desativar instalação do projeto

[tool.poetry.dependencies]
python = "^3.12.10"
django = "6.0"
djangorestframework = "3.16.1"
psycopg2-binary = "^2.9.11"
djangorestframework-simplejwt = "^5.5.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
black = "^24.3"
isort = "^6.0"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
flake8 = "^6.1"
